requirements.txt
supervision
lightly-train
Pillow
empy==3.3.4
pyros-genmsg 
setuptools==65.5.1
catkin_pkg
lark
colcon-common-extensions
geographiclib
numpy<2,>=1.23
